chore: rename cmd.py to main.py and cli_wrapper() to main()
And fix the code in `if __name__ == "__main__"`.
