+++ tests/cmd/test_crash_log.py
@@ -4,7 +4,7 @@ from unittest import mock
 import click.exceptions
 import pytest
 
-from ggshield.cmd.cmd import cli_wrapper
+from ggshield.cmd.main import main
 
 
 CRASH_LOG_ENV = {"GITGUARDIAN_CRASH_LOG": "true"}
@@ -12,11 +12,11 @@ CRASH_LOG_ENV = {"GITGUARDIAN_CRASH_LOG": "true"}
 
 def test_syntax_error_no_crash_log():
     with pytest.raises(SystemExit) as e:
-        cli_wrapper(args=["foo"])
+        main(args=["foo"])
     assert e.value.code != 0
 
 
 @mock.patch.dict(os.environ, CRASH_LOG_ENV)
 def test_syntax_error_crash_log():
     with pytest.raises(click.exceptions.UsageError):
-        cli_wrapper(args=["foo"])
+        main(args=["foo"])
